Restrict the path of iiidevops_install.pl and update-perl.pl execution

+++ bin/update-perl.pl
@@ -15,14 +15,21 @@ if ($cmd_msg ne 'rkeuser') {
 	exit;
 }
 
+# check running path
+$home_path = '/home/rkeuser';
+if ($Bin ne $home_path) {
+	log_print("You must run the installation script in [$home_path]!\n");
+	exit;
+}
+
 # Install iiidevops Deploy Scripts
 $cmd = <<END;
-cd ~;
+cd $home_path;
 wget -O $ins_repo.zip https://github.com/iii-org/deploy-devops/archive/$ins_repo.zip
 unzip -o $ins_repo.zip;
 rm -rf deploy-devops;
 mv deploy-devops-$ins_repo deploy-devops;
-find ~/deploy-devops -type f -name \"*.pl\" -exec chmod a+x {} \\;
+find $home_path/deploy-devops -type f -name \"*.pl\" -exec chmod a+x {} \\;
 END
 print("Install iiidevops Deploy Scripts..\n");
 system($cmd);
